Placed IGCMenu.swift at project level folder
